Access Control List

+++ api/routes/users.js
@@ -3,6 +3,7 @@ const usersModel = require('../models/users');
 const auth = require('../controllers/auth');
 const router = express.Router()
 const {validateUser, validateUserUpdate} = require('../controllers/validation');
+const can = require('../permissions/users');
 
 router.use(express.json());
 
@@ -13,23 +14,24 @@ router.use(function timeLog (req, res, next) {
 
 const createUser = async (req, res) => {
     let result = await usersModel.addUser(req.body);
-    if (Array.isArray(result)) {
-        result[0].password = '';
-        return res.status(200).json(result);
-    } else if (result.severity === 'ERROR') {
-        return res.status(500).json('User Exist or BAD REQUEST');
-    }
+    console.log(result);
+    return res.status(200).json('User created succesfully');
 }
 
 const getAll = async (req, res) => {
-    let result = await usersModel.getAll().catch((err) => console.log(err));
-    if (Array.isArray(result) && result.length) {
-        result[0].password = '';
-        return res.status(200).json(result);
-    } else if (Array.isArray(result) && result.length === 0) {
-        return res.status(404).json('No user found');
+    const permission = can.readAll(req.user);
+    console.log(permission);
+    if (!permission.granted) {
+        res.status(403).json("You don't have permission");
     } else {
-        return res.status(500).json('Error found when getting users');
+        let result = await usersModel.getAll().catch((err) => console.log(err));
+        if (Array.isArray(result) && result.length) {
+            return res.status(200).json(result);
+        } else if (Array.isArray(result) && result.length === 0) {
+            return res.status(404).json('No user found');
+        } else {
+            return res.status(500).json('Error found when getting users');
+        }
     }
 }
 
@@ -38,8 +40,13 @@ const getById = async (req, res) => {
     let result = await usersModel.getById(id).catch((err) => console.log(err));
 
     if (Array.isArray(result) && result.length) {
-        result[0].password = '';
-        return res.status(200).json(result);
+        const data = result[0];
+        let permission = can.read(req.user, data);
+        if (!permission.granted) {
+            res.status(403).json("You don't have permission");
+        } else {
+            return res.status(200).json(result);
+        } 
     } else if (Array.isArray(result) && result.length === 0) {
         return res.status(404).json('No user found');
     } else {
@@ -51,8 +58,13 @@ const updateUser = async (req, res) => {
     let { id } = req.params;
     let result = await usersModel.updateUser(req.body, id).catch((err) => console.log(err));
     if (Array.isArray(result) && result.length) {
-        result[0].password = '';
-        return res.status(200).json(result);
+        let data = result[0];
+        const permission = can.update(req.user, data);
+        if (!permission.granted) {
+            res.status(403).json("You don't have permission");
+        } else {
+            return res.status(200).json(result);
+        }
     } else if (result === undefined) {
         return res.status(500).json('Error when updating the user');
     } else if (Array.isArray(result) && result.length === 0) {
@@ -62,13 +74,23 @@ const updateUser = async (req, res) => {
 
 const deleteUser = async (req, res) => {
     let { id } = req.params;
-    let result = await usersModel.deleteUser(id).catch((err) => console.log(err));
-    if (result) {
-        return res.status(200).json('User deleted succesfully');
-    } else if (result === undefined) {
-        return res.status(500).json('Error when deleting a user');
-    } else {
-        return res.status(400).json('No user found to delete');
+    let result = await usersModel.getById(id);
+    if (result.length) {
+        const data = result[0];
+        console.log('trying to delete', data);
+        const permission = can.delete(req.user, data);
+        if (!permission.granted) {
+            res.status(403).json("You don't have permission");
+        } else {
+            let result = await usersModel.deleteUser(id).catch((err) => console.log(err));
+            if (result) {
+                return res.status(200).json('User deleted succesfully');
+            } else if (result === undefined) {
+                return res.status(500).json('Error when deleting a user');
+            } else {
+                return res.status(400).json('No user found to delete');
+            }
+        }
     }
 }
